refactor: add resoures directory
