+++ package.json
@@ -1,75 +1,76 @@
 {
-    "name": "tdesign-icons",
-    "version": "0.0.0-NOT-VALID",
-    "description": "TDesign Icon 图标统一管理仓库",
-    "repository": {
-        "type": "git",
-        "url": "git@github.com:Tencent/tdesign-icons.git"
-    },
-    "keywords": [
-        "tdesign",
-        "icon",
-        "tdesign-icons-react",
-        "tdesign-icons-vue",
-        "tdesign-icons-vue-next",
-        "tdesign-icons-view",
-        "tdesign-icons-svg"
+  "name": "tdesign-icons",
+  "private": true,
+  "description": "TDesign Icon 图标统一管理仓库",
+  "workspaces": [
+    "packages/**"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:Tencent/tdesign-icons.git"
+  },
+  "keywords": [
+    "tdesign",
+    "icon"
+  ],
+  "author": "tdesign",
+  "lint-staged": {
+    "packages/**/{src,gulp}/**/*.{js,ts,tsx,jsx,vue}": [
+      "eslint --fix"
     ],
-    "author": "tdesign",
-    "lint-staged": {
-        "packages/**/{src,gulp}/**/*.{js,ts,tsx,jsx,vue}": [
-            "eslint --fix"
-        ],
-        "gulp/**/*.{js,ts,tsx,jsx,vue}": [
-            "eslint --fix"
-        ]
-    },
-    "scripts": {
-        "husky": "husky",
-        "prepare": "husky install",
-        "lint": "eslint ./packages --max-warnings 0 --ext .ts,.tsx --fix",
-        "generate": "cross-env gulp --require ts-node/register/transpile-only"
-    },
-    "devDependencies": {
-        "@types/fs-extra": "^9.0.9",
-        "@types/gulp": "^4.0.6",
-        "@types/gulp-concat": "^0.0.32",
-        "@types/gulp-rename": "^0.0.33",
-        "@types/gulp-sourcemaps": "^0.0.34",
-        "@types/gulp-svg-sprite": "^1.2.32",
-        "@types/merge-stream": "^1.1.2",
-        "@types/merge2": "^1.3.0",
-        "@types/svg-parser": "^2.0.1",
-        "@types/svgo": "^1.3.3",
-        "@types/through2": "^2.0.36",
-        "@typescript-eslint/eslint-plugin": "^4.20.0",
-        "@typescript-eslint/parser": "^4.20.0",
-        "camelcase": "^6.2.0",
-        "cross-env": "^7.0.3",
-        "decamelize": "^5.0.0",
-        "del": "^6.0.0",
-        "eslint": "^7.23.0",
-        "eslint-config-airbnb-base": "^14.2.1",
-        "eslint-plugin-import": "^2.25.2",
-        "fs-extra": "^9.1.0",
-        "gulp": "^4.0.2",
-        "gulp-clone": "^2.0.1",
-        "gulp-concat": "^2.6.1",
-        "gulp-header": "^2.0.9",
-        "gulp-iconfont": "10.0.3",
-        "gulp-iconfont-css": "^3.0.0",
-        "gulp-merge": "^0.1.1",
-        "gulp-rename": "^2.0.0",
-        "gulp-sourcemaps": "^3.0.0",
-        "gulp-svg-sprite": "^1.5.0",
-        "gulp-typescript": "^6.0.0-alpha.1",
-        "husky": "^6.0.0",
-        "lint-staged": "^10.5.4",
-        "merge-stream": "^2.0.0",
-        "merge2": "^1.4.1",
-        "svg-parser": "^2.0.4",
-        "svgo": "^1.3.2",
-        "ts-node": "^9.1.1",
-        "typescript": "^4.1.3"
-    }
+    "gulp/**/*.{js,ts,tsx,jsx,vue}": [
+      "eslint --fix"
+    ]
+  },
+  "scripts": {
+    "husky": "husky",
+    "prepare": "husky install",
+    "lint": "eslint ./packages --max-warnings 0 --ext .ts,.tsx --fix",
+    "generate": "cross-env gulp --require ts-node/register/transpile-only"
+  },
+  "devDependencies": {
+    "@types/fs-extra": "^9.0.9",
+    "@types/gulp": "^4.0.6",
+    "@types/gulp-concat": "^0.0.32",
+    "@types/gulp-rename": "^0.0.33",
+    "@types/gulp-sourcemaps": "^0.0.34",
+    "@types/gulp-svg-sprite": "^1.2.32",
+    "@types/merge-stream": "^1.1.2",
+    "@types/merge2": "^1.3.0",
+    "@types/svg-parser": "^2.0.1",
+    "@types/svgo": "^1.3.3",
+    "@types/through2": "^2.0.36",
+    "@typescript-eslint/eslint-plugin": "^4.20.0",
+    "@typescript-eslint/parser": "^4.20.0",
+    "camelcase": "^6.2.0",
+    "cross-env": "^7.0.3",
+    "decamelize": "^5.0.0",
+    "del": "^6.0.0",
+    "eslint": "^7.23.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.25.2",
+    "fs-extra": "^9.1.0",
+    "gulp": "^4.0.2",
+    "gulp-clone": "^2.0.1",
+    "gulp-concat": "^2.6.1",
+    "gulp-header": "^2.0.9",
+    "gulp-iconfont": "10.0.3",
+    "gulp-iconfont-css": "^3.0.0",
+    "gulp-merge": "^0.1.1",
+    "gulp-rename": "^2.0.0",
+    "gulp-sourcemaps": "^3.0.0",
+    "gulp-svg-sprite": "^1.5.0",
+    "gulp-typescript": "^6.0.0-alpha.1",
+    "husky": "^6.0.0",
+    "lint-staged": "^10.5.4",
+    "merge-stream": "^2.0.0",
+    "merge2": "^1.4.1",
+    "svg-parser": "^2.0.4",
+    "svgo": "^1.3.2",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
+  },
+  "dependencies": {
+    "lerna": "^6.6.1"
+  }
 }
