+++ resources/icon-font/.npmignore
