+++ resources/svg-sprite/.gitignore
