+++ resources/svg-sprite/.npmignore
