+++ resources/svg-sprite/gulp/index.ts
