+++ resources/svg-sprite/gulp/template/index.js
