+++ resources/svg-sprite/gulp/use-svg-sprite-template.ts
