+++ resources/svg-sprite/package.json
